update to highlighter code split
