+++ packages/lexical-code/src/index.ts
@@ -1,22 +1,14 @@
-import {
-  $createCodeNode,
-  $isCodeNode,
-  CodeNode,
-  registerCodeHighlighting,
-} from './CodeHighlighter';
-import {
-  getCodeLanguages,
-  getDefaultCodeLanguage,
-  getEndOfCodeInLine,
-  getFirstCodeHighlightNodeOfLine,
-  getLastCodeHighlightNodeOfLine,
-  getStartOfCodeInLine,
-  registerCodeIndent,
-} from './EditorShortcuts';
+import {registerCodeHighlighting} from './CodeHighlighter';
 import {
   $createCodeHighlightNode,
   $isCodeHighlightNode,
   CodeHighlightNode,
+} from './CodeHighlightNode';
+import {$createCodeNode, $isCodeNode, CodeNode} from './CodeNode';
+import {registerCodeIndent} from './EditorShortcuts';
+import {
+  getFirstCodeHighlightNodeOfLine,
+  getLastCodeHighlightNodeOfLine,
 } from './HighlighterHelper';
 
 export {
@@ -26,12 +18,8 @@ export {
   $isCodeNode,
   CodeHighlightNode,
   CodeNode,
-  getCodeLanguages,
-  getDefaultCodeLanguage,
-  getEndOfCodeInLine,
   getFirstCodeHighlightNodeOfLine,
   getLastCodeHighlightNodeOfLine,
-  getStartOfCodeInLine,
   registerCodeHighlighting,
   registerCodeIndent,
 };
